refactor(image): move directory
