--- extractor/testdata/test_index_json_dir.oci/oci-layout
@@ -1,1 +1,0 @@
-{"imageLayoutVersion": "1.0.0"}
