--- extractor/testdata/test_index_json_dir.oci/test_bad_index_json.oci/index.json
@@ -1,1 +1,0 @@
-foobar
