--- extractor/testdata/test_invalid_oci_image.oci/index.json
@@ -1,1 +1,0 @@
-{"schemaVersion":2,"manifests":[{"mediaType":"application/badmedia+json","digest":"sha256:bad744871f99e0ff8e6f253244836ed34c5d805fdb096d3a205ffaf5e9073cab","size":345}]}
