--- extractor/testdata/test_invalid_oci_image.oci/oci-layout
@@ -1,1 +1,0 @@
-{"imageLayoutVersion": "1.0.0"}
