--- extractor/testdata/test_no_valid_manifests.oci/index.json
@@ -1,1 +1,0 @@
-{}
