--- extractor/testdata/test_no_valid_manifests.oci/oci-layout
@@ -1,1 +1,0 @@
-{"imageLayoutVersion": "1.0.0"}
